+++ ui.frontend/src/providers/ResultsProvider.tsx
@@ -5,7 +5,6 @@ import {
   PropsWithChildren,
   Dispatch,
   useCallback,
-  useRef,
   useMemo,
   useEffect,
 } from 'react';
@@ -14,6 +13,8 @@ import { useAlertDispatcher } from './WizardAlertProvider';
 import { useLogger } from './LoggingProvider';
 import exportFromJSON from 'export-from-json';
 import { ExportType } from 'export-from-json/src/types';
+import { useRenderCount } from 'src/utility';
+import { PackagingProvider } from './PackagingProvider';
 
 /**
  * This is an extension of the OOTB export-from-json types,
@@ -61,8 +62,8 @@ const ResultsDispatchContext = createContext<Dispatch<ResultsDispatchProps>>(nul
  */
 export function ResultsProvider({ children }: PropsWithChildren) {
   const logger = useLogger();
-  const renderCount = useRef(0);
-  logger.debug({ message: `ResultsProvider[${++renderCount.current}] render()` });
+  const renderCount = useRenderCount();
+  logger.debug({ message: `ResultsProvider[${renderCount}] render()` });
   const alertDispatcher = useAlertDispatcher();
   // non-modified results is more or less used as a reference for rendering logic
   const [results, setResults] = useState([] as Result[]);
@@ -75,7 +76,7 @@ export function ResultsProvider({ children }: PropsWithChildren) {
 
   const updateResults = useCallback(
     ({ results, caller }: ResultsDispatchProps) => {
-      logger.debug({ message: `ResultsProvider[${renderCount.current}] results dispatcher called by `, caller });
+      logger.debug({ message: `ResultsProvider results dispatcher called by `, caller });
       if (!results || results.length < 1) {
         alertDispatcher({
           message: 'No results were found.',
@@ -98,11 +99,11 @@ export function ResultsProvider({ children }: PropsWithChildren) {
   useEffect(() => {
     let arr = results;
     if (filter) {
-      logger.debug({ message: `ResultsProvider[${renderCount.current}] useEffect.filter()` });
+      logger.debug({ message: `ResultsProvider useEffect.filter()` });
       arr = filterResults(arr, filter);
     }
     if (orderBy) {
-      logger.debug({ message: `ResultsProvider[${renderCount.current}] useEffect.sort()` });
+      logger.debug({ message: `ResultsProvider useEffect.sort()` });
       arr = stableSort(arr, getComparator(order, orderBy));
     }
     setTableResults(arr);
@@ -146,7 +147,9 @@ export function ResultsProvider({ children }: PropsWithChildren) {
 
   return (
     <ResultsContext.Provider value={value}>
-      <ResultsDispatchContext.Provider value={updateResults}>{children}</ResultsDispatchContext.Provider>
+      <ResultsDispatchContext.Provider value={updateResults}>
+        <PackagingProvider>{children}</PackagingProvider>
+      </ResultsDispatchContext.Provider>
     </ResultsContext.Provider>
   );
 }
@@ -155,7 +158,7 @@ export function useResults() {
   return useContext(ResultsContext);
 }
 
-export function useResultsDispatch() {
+export function useResultsDispatcher() {
   return useContext(ResultsDispatchContext);
 }