+++ ui.frontend/src/utility/index.ts
@@ -3,6 +3,7 @@
  * Third-party dependencies may still be imported, but to avoid circular dependencies, do not use other modules.
  * Exposed via 'src/utility'
  */
+export * from './hooks';
 export * from './csrf';
 export * from './libs';
 export * from './createFetcher';